move mapping to props
