+++ packages/frontend/src/components/project/PermissionsSection.tsx
@@ -6,7 +6,7 @@ import { ContractEntry, TechnologyContract } from './ContractEntry'
 import { Section } from './Section'
 
 export interface PermissionsSectionProps {
-  permissions: ProjectPermission[]
+  permissions: TechnologyContract[]
   verificationStatus: VerificationStatus
 }
 
@@ -20,10 +20,10 @@ export function PermissionsSection({
         The system uses the following set of permissioned addresses:
       </h3>
       <div className="flex flex-wrap gap-4 my-4">
-        {permissions.map(toContract).map((contract, i) => (
+        {permissions.map((permission, i) => (
           <React.Fragment key={i}>
             <ContractEntry
-              contract={contract}
+              contract={permission}
               verificationStatus={verificationStatus}
             />
           </React.Fragment>
@@ -31,19 +31,4 @@ export function PermissionsSection({
       </div>
     </Section>
   )
-}
-
-function toContract(permission: ProjectPermission): TechnologyContract {
-  const links = permission.accounts.slice(1).map((account) => ({
-    name: `${account.address.slice(0, 6)}…${account.address.slice(38, 42)}`,
-    address: account.address,
-    href: `https://etherscan.io/address/${account.address}#code`,
-  }))
-
-  return {
-    name: permission.name,
-    address: permission.accounts[0]?.address,
-    description: permission.description,
-    links,
-  }
 }
